Created a new class LinkEmbedding_Int that holds the code common to all LinkEmbeddingX classes that deal with integer coordinates. (The name of this class might be changed to something nice later.) LinkEmbedding2/3/4/4a are specialized type aliases with Proresector2/3/4/4a as backend.

+++ src/LinkEmbedding_Int/FromFile.hpp
@@ -26,7 +26,10 @@ static LinkEmbedding_T FromInString( mref<Tools::InString> s )
     component_ptr_agg.push_back(Size_T(0));
     std::vector<Int> color_agg;
     
-    auto lc = [&component_ptr_agg](){ return component_ptr_agg.size() - Size_T(1); };
+    auto lc = [&component_ptr_agg]()
+    {
+        return static_cast<Int>(component_ptr_agg.size() - Size_T(1));
+    };
     
     bool color_declaredQ    = false; // Whether some #color attribute has beend found before.
     bool comp_wo_colorQ     = false; // Whether there is some component without #color.
